Implemented file checking and recovery options for all database files.
Implemented new Admin login panel to add the first account to the security database file.

Set the keyfile as hidden in the file system.

Fixed a bug when setting default JButton on the login panel Enter would crash on any other panel. Set the default JButton to null on all panel constructors that do not use a default button.